+++ StripePayments/StripePayments/Source/PaymentHandler/STPPaymentHandler.swift
@@ -980,12 +980,7 @@ public class STPPaymentHandler: NSObject {
                 if lastSetupError.code == STPSetupIntentLastSetupError.CodeAuthenticationFailure {
                     action.complete(
                         with: STPPaymentHandlerActionStatus.failed,
-                        error: _error(
-                            for: .notAuthenticatedErrorCode,
-                            apiErrorCode: lastSetupError.code,
-                            declineCode: lastSetupError.declineCode,
-                            apiErrorType: lastSetupError.allResponseFields["type"] as? String
-                        )
+                        error: _error(for: .notAuthenticatedErrorCode)
                     )
                 } else if lastSetupError.type == .card {
                     action.complete(
@@ -993,20 +988,13 @@ public class STPPaymentHandler: NSObject {
                         error: _error(
                             for: .paymentErrorCode,
                             apiErrorCode: lastSetupError.code,
-                            localizedDescription: lastSetupError.message,
-                            declineCode: lastSetupError.declineCode,
-                            apiErrorType: lastSetupError.allResponseFields["type"] as? String
+                            localizedDescription: lastSetupError.message
                         )
                     )
                 } else {
                     action.complete(
                         with: STPPaymentHandlerActionStatus.failed,
-                        error: _error(
-                            for: .paymentErrorCode,
-                            apiErrorCode: lastSetupError.code,
-                            declineCode: lastSetupError.declineCode,
-                            apiErrorType: lastSetupError.allResponseFields["type"] as? String
-                        )
+                        error: _error(for: .paymentErrorCode, apiErrorCode: lastSetupError.code)
                     )
                 }
             } else {
@@ -1063,12 +1051,7 @@ public class STPPaymentHandler: NSObject {
                 {
                     action.complete(
                         with: STPPaymentHandlerActionStatus.failed,
-                        error: _error(
-                            for: .notAuthenticatedErrorCode,
-                            apiErrorCode: lastPaymentError.code,
-                            declineCode: lastPaymentError.declineCode,
-                            apiErrorType: lastPaymentError.allResponseFields["type"] as? String
-                        )
+                        error: _error(for: .notAuthenticatedErrorCode)
                     )
                 } else if lastPaymentError.type == .card {
 
@@ -1077,20 +1060,13 @@ public class STPPaymentHandler: NSObject {
                         error: _error(
                             for: .paymentErrorCode,
                             apiErrorCode: lastPaymentError.code,
-                            localizedDescription: lastPaymentError.message,
-                            declineCode: lastPaymentError.declineCode,
-                            apiErrorType: lastPaymentError.allResponseFields["type"] as? String
+                            localizedDescription: lastPaymentError.message
                         )
                     )
                 } else {
                     action.complete(
                         with: STPPaymentHandlerActionStatus.failed,
-                        error: _error(
-                            for: .paymentErrorCode,
-                            apiErrorCode: lastPaymentError.code,
-                            declineCode: lastPaymentError.declineCode,
-                            apiErrorType: lastPaymentError.allResponseFields["type"] as? String
-                        )
+                        error: _error(for: .paymentErrorCode, apiErrorCode: lastPaymentError.code)
                     )
                 }
             } else {
@@ -2351,16 +2327,11 @@ public class STPPaymentHandler: NSObject {
         for errorCode: STPPaymentHandlerErrorCode,
         apiErrorCode: String? = nil,
         loggingSafeErrorMessage: String? = nil,
-        localizedDescription: String? = nil,
-        declineCode: String? = nil,
-        apiErrorType: String? = nil
+        localizedDescription: String? = nil
     ) -> NSError {
         var userInfo = [String: String]()
         userInfo[STPError.errorMessageKey] = loggingSafeErrorMessage
         userInfo[NSLocalizedDescriptionKey] = localizedDescription
-        userInfo[STPError.stripeErrorCodeKey] = apiErrorCode
-        userInfo[STPError.stripeDeclineCodeKey] = declineCode
-        userInfo[STPError.stripeErrorTypeKey] = apiErrorType
         switch errorCode {
         // 3DS(2) flow expected user errors
         case .notAuthenticatedErrorCode: